requirements.txt
google-generativeai>=0.3.0
PyYAML>=6.0
feedparser>=6.0
requests>=2.32
python-dateutil>=2.9
beautifulsoup4>=4.12